requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
langdetect>=1.0.9
Pillow>=10.0.0
pytesseract>=0.3.10
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
googletrans>=4.0.0
pydantic>=2.0.0